requirements.txt
pybaseball
pandas
numpy
sqlalchemy
python-dotenv
psycopg2-binary
scikit-learn
jupyter
pyarrow
